+++ mm/kasan/CMakeLists.txt
@@ -19,13 +19,13 @@
 # the License.
 #
 # ##############################################################################
+set(SRCS hook.c)
+
 if(CONFIG_MM_KASAN)
-  target_sources(mm PRIVATE kasan.c)
-
-  target_compile_options(mm PRIVATE -fno-sanitize=kernel-address)
-
-  if(NOT CONFIG_LTO_NONE)
-    target_compile_options(mm PRIVATE -fno-lto)
-  endif()
-
+  list(APPEND SRCS generic.c)
+  set_source_files_properties(generic.c PROPERTIES COMPILE_FLAGS
+                                                   -fno-sanitize=kernel-address)
+  set_source_files_properties(generic.c PROPERTIES COMPILE_FLAGS -fno-lto)
 endif()
+
+target_sources(mm PRIVATE ${SRCS})
