+++ mm/kasan/kasan.h
@@ -27,6 +27,7 @@
  * Included Files
  ****************************************************************************/
 
+#include <stdbool.h>
 #include <stddef.h>
 
 /****************************************************************************
@@ -34,11 +35,12 @@
  ****************************************************************************/
 
 #ifndef CONFIG_MM_KASAN
+#  define kasan_is_poisoned(addr, size) false
 #  define kasan_poison(addr, size)
 #  define kasan_unpoison(addr, size)
 #  define kasan_register(addr, size)
 #  define kasan_init_early()
-#endif
+#else
 
 /****************************************************************************
  * Public Function Prototypes
@@ -52,7 +54,22 @@ extern "C"
 #define EXTERN extern
 #endif
 
-#ifdef CONFIG_MM_KASAN
+/****************************************************************************
+ * Name: kasan_is_poisoned
+ *
+ * Description:
+ *   Check if the memory range is poisoned
+ *
+ * Input Parameters:
+ *   addr - range start address
+ *   size - range size
+ *
+ * Returned Value:
+ *   true if the memory range is poisoned, false otherwise.
+ *
+ ****************************************************************************/
+
+bool kasan_is_poisoned(FAR const void *addr, size_t size);
 
 /****************************************************************************
  * Name: kasan_poison
@@ -126,11 +143,11 @@ void kasan_register(FAR void *addr, FAR size_t *size);
 
 void kasan_init_early(void);
 
-#endif /* CONFIG_MM_KASAN */
-
 #undef EXTERN
 #ifdef __cplusplus
 }
 #endif
+
+#endif /* CONFIG_MM_KASAN */
 
 #endif /* __MM_KASAN_KASAN_H */
